# getRenderableIndexes and cacheFromPhysicalToVisualIndexes optimization