### Context
Initialization and `updateSettings` are slow when we have 50k indexes. Investigating 🕵️ 
An example from #3770: https://jsfiddle.net/AMBudnik/6y50qmh9/

* https://github.com/handsontable/handsontable/commit/272a22fde48062ca64fbf67828f7f8df8d0ca019 `getRenderableIndexes` iterations loop grows with the number of indexes, because `includes` has more and more indexes to iterate over. Up to `N = length` operations for each row, `N * N` operations in the worst-case scenario. We can switch to O(1) op here and have a stable `N` number of operations
* https://github.com/handsontable/handsontable/pull/6936/commits/90abc5263aac68b99ac2b9feab969806c0925d1c `cacheFromPhysicalToVisualIndexes` iteration loop grows with the number of indexes, because `indexOf` has more and more indexes to scan over. In the worst-case scenario it's `N*N` as above. `notTrimmedIndexes` are already in the right order (build with `indexesSequence` and cached) so we can use that to just rewrite indexes to cache with `N` interactions only. 
* Same commit ☝️ Use trimmed indexes instead of the whole collection. `N` iterations becomes even smaller, reduced to the trimmed indexes which we previously filtered out. It should be faster when we have multiple indexes trimmed.
* https://github.com/handsontable/handsontable/pull/6936/commits/db00123e043cc4239e3faeeaa8e723e0754232ad Do not recreate filtered renderable collection, we already have it cached
* https://github.com/handsontable/handsontable/pull/6936/commits/2580c2e50bb27442771d07342be75a062b9b9584 iterate only over renderable indexes, this should be quicker when we have multiple indexes hidden
* https://github.com/handsontable/handsontable/pull/6936/commits/f6ac09ed694dee5dac669439b0da26cd5025d3de might be controversial, I know there were some discussions helpers vs native. We were using both in this file


### How has this been tested?
I'm running tests assuming refactor changes just op time. 
Performance compared in Chrome perf tools.

#### Results
https://github.com/handsontable/handsontable/commit/272a22fde48062ca64fbf67828f7f8df8d0ca019 reduced the time for `getRenderableIndexes` in the given example from 2s to 2.5ms
https://github.com/handsontable/handsontable/pull/6936/commits/90abc5263aac68b99ac2b9feab969806c0925d1c reduced the time for `cacheFromPhysicalToVisualIndexes` in the given example from 2.1s to 5.4ms

Other commits are not that impressive but should be noticeable when we have multiple indexes trimmed or hidden or both.

Total `initToLength` from 4.22s to 29ms.

### Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature or improvement (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [ ] Additional language file or change to the existing one (translations)

### Related issue(s):
1. #3770
2.
3.

### Checklist:
<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the code style of this project,
- [ ] My change requires a change to the documentation.