@wojciechczerniak 2020-05-27
Changes this:
![Screenshot 2020-05-27 at 02 11 32](https://user-images.githubusercontent.com/4718998/82962107-88bed880-9fbf-11ea-8565-f839dbab019d.png)

Into this:
![Screenshot 2020-05-27 at 02 12 13](https://user-images.githubusercontent.com/4718998/82962111-8fe5e680-9fbf-11ea-827c-51821f73443c.png)

There the right side has a lot of calls to getCellMeta, may be improved with #6930 by @budnix 

---

@budnix Maybe you can take a look as a reviewer, since you're close to the perf subject this week